+++ extension/Cargo.toml
@@ -26,6 +26,7 @@ serde_json = "1.0"
 reqwest = { version = "0.11", features = ["json"] }
 tokio = { version = "1", features = ["full"] }
 uuid = { version = "1.1", features = ["v4", "v5", "serde"] }
+anyhow = "1.0"
 
 [dev-dependencies]
 pgrx-tests = "=0.11.4"
