+++ extension/src/controller/mod.rs
@@ -1,1 +1,2 @@
+pub mod bgworker;
 pub mod dv_builder;
