+++ extension/src/model/mod.rs
@@ -1,2 +1,3 @@
 pub mod source_objects;
 pub mod dv_transformer_schema;
+pub mod queries;
