+++ extension/src/utility/mod.rs
@@ -1,0 +1,3 @@
+pub mod ollama_client;
+pub mod setup;
+pub mod guc;
