+++ extension/src/utility/ollama_client.rs
