+++ extension/src/utility/setup.rs
@@ -1,8 +1,8 @@
 use pgrx::prelude::*;
 
-extension_sql_file!("../sql/sample_source_tables/seller.sql");
-extension_sql_file!("../sql/sample_source_tables/customer.sql");
-extension_sql_file!("../sql/info_tables.sql");
+extension_sql_file!("sql/sample_source_tables/seller.sql");
+extension_sql_file!("sql/sample_source_tables/customer.sql");
+extension_sql_file!("sql/info_tables.sql");
 
 
 // Creating Extension TABLES ** Purely ** as a Hello World + Example.
